fix: rename 'builtin' logger to 'default', put it in its own package
Probably very few users will want to reference the default (nee "builtin")
logger directly, so this commit moves it to a sub-package with a very
clear name.

We also move the "Level" constants into that package, because they
are only relevant for that logger.  Anyone who is coming to us with
an existing logging framework will have there own LogLevel constants
that are defined outside of Momento.
