+++ config/logger/momento-logger.go
@@ -1,15 +1,5 @@
 package logger
 
-type LogLevel int
-
-const (
-	TRACE LogLevel = 5
-	DEBUG LogLevel = 10
-	INFO  LogLevel = 20
-	WARN  LogLevel = 30
-	ERROR LogLevel = 40
-)
-
 type MomentoLogger interface {
 	Trace(message string, args ...string)
 	Debug(message string, args ...string)
